+++ db/migrate/039_add_index_jockey_id_and_race_id_to_entries.rb
@@ -1,0 +1,5 @@
+class AddIndexJockeyIdAndRaceIdToEntries < ActiveRecord::Migration[4.2]
+  def change
+    add_index :entries, %i[jockey_id race_id]
+  end
+end