+++ extract.rb
@@ -1,3 +1,5 @@
+require 'active_support'
+require 'active_support/core_ext/object/blank'
 require 'nokogiri'
 require_relative 'config/initialize'
 require_relative 'db/connect'
@@ -7,37 +9,57 @@ Dir['models/*.rb'].each {|f| require_relative f }
 class Extractor
   Dir['extract/*'].each {|f| require_relative f }
   BACKUP_DIR = File.join(APPLICATION_ROOT, 'backup')
+  VALID_OPERATIONS = %w[create update upsert].freeze
+  DEFAULT_OPERATION = 'create'.freeze
 
   include ArgumentUtil
 
-  def self.work!
-    logger = DenebolaLogger.new(Settings.logger.path.extract)
-    ArgumentUtil.logger = logger
+  class << self
+    def work!
+      logger = DenebolaLogger.new(Settings.logger.path.extract)
+      ArgumentUtil.logger = logger
+      ApplicationRecord.operation = operation
 
-    check_operation(operation)
-    ApplicationRecord.operation = operation
+      extractor = new(logger)
 
-    extractor = new(logger)
+      (from..to).each do |date|
+        race_ids = extractor.fetch_race_ids(date)
 
+        Array.wrap(race_ids).each do |race_id|
+          race_html = extractor.fetch_race(race_id)
+          race_html = Nokogiri::HTML.parse(race_html)
+          next if race_html.nil?
 
-    (from..to).each do |date|
-      race_ids = extractor.fetch_race_ids(date)
+          race = extractor.update_race_info(race_html, race_id)
+          next if race.nil?
 
-      Array.wrap(race_ids).each do |race_id|
-        race_html = extractor.fetch_race(race_id)
-        race_html = Nokogiri::HTML.parse(race_html)
-        next if race_html.nil?
+          _, *rows =
+            race_html.xpath('//table[contains(@class, "race_table")]').search('tr')
+          rows.each {|row| extractor.update_entry_info(row, race_id) rescue next }
 
-        race = extractor.update_race_info(race_html, race_id)
-        next if race.nil?
-
-        _, *rows =
-          race_html.xpath('//table[contains(@class, "race_table")]').search('tr')
-        rows.each {|row| extractor.update_entry_info(row, race_id) rescue next }
-
-        payoff_attribute = extract_payoff(race_html)
-        race.create_or_update_payoff(payoff_attribute.compact)
+          payoff_attribute = extract_payoff(race_html)
+          race.create_or_update_payoff(payoff_attribute.compact)
+        end
       end
+    end
+
+    def from
+      super.blank? ? Date.today - 30 : Date.parse(super)
+    end
+
+    def to
+      super.blank? ? Date.today : Date.parse(super)
+    end
+
+    def operation
+      operation = super.blank? ? DEFAULT_OPERATION : super
+
+      unless VALID_OPERATIONS.include?(operation)
+        logger.error("invalid operation specified: #{operation}")
+        raise StandardError
+      end
+
+      operation
     end
   end
 
@@ -82,13 +104,14 @@ class Extractor
   end
 
   def update_entry_info(html, race_id)
-    attribute = extract_entry(html).merge(race_id: race_id)
+    race = Race.find_by(race_id: race_id)
+    attribute = extract_entry(html).merge(race_id: race.id)
     log_attribute = Entry.log_attribute.merge(attribute.slice(:race_id, :number))
     entry = handle_active_record_error(log_attribute) do
       Entry.create_or_update!(attribute)
     end
-    entry.race = Race.find_by(race_id: race_id)
-    jockey = update_jockey_info(attribute[:jockey_id], race_id)
+    entry.race = race
+    jockey = update_jockey_info(attribute[:jockey_id], race.race_id)
 
     if jockey and entry
       jockey.results << entry