+++ models/horse.rb
@@ -25,35 +25,46 @@ class Horse < ApplicationRecord
     super.merge(resource: 'horse')
   end
 
-  def average_prize_money(time)
-    results_before(time).map(&:prize_money).inject(:+) / entry_times(time).to_f
+  def extra_attribute(time)
+    results_before = results.joins(:race)
+                            .where('races.start_time < ?', time)
+                            .order('races.start_time desc')
+
+    blank = if results_before.second
+              (time.to_date - results_before.first.race.start_time.to_date).to_i
+            else
+              0
+            end
+
+    {
+      blank: blank,
+      entry_times: results_before.size,
+      horse_average_prize_money: average_prize_money(results_before),
+      last_race_order: last_race_order(results_before),
+      rate_within_third: rate_within_third(results_before),
+      second_last_race_order: second_last_race_order(results_before),
+      win_times: results_before.count(&:won),
+    }
   end
 
-  def entry_times(time)
-    results_before(time).size
+  def average_prize_money(results)
+    results.average(:prize_money) || 0
   end
 
-  def last_race_order(time)
-    results_before(time).second&.order&.to_i || 0
+  def last_race_order(results)
+    results.first&.order&.to_i || 0
   end
 
-  def rate_within_third(time)
-    within_third = results_before(time).first(3).count do |result|
+  def rate_within_third(results)
+    return 0 if results.empty?
+
+    within_third = results.first(3).count do |result|
       %w[1 2 3].include?(result.order)
     end
-    within_third / 3.0
+    within_third / [3.0, results.size.to_f].min
   end
 
-  def second_last_race_order(time)
-    results_before(time).third&.order&.to_i || 0
-  end
-
-  def win_times(time)
-    results_before(time).count(&:won)
-  end
-
-  def results_before(time)
-    @results_before ||= results.joins(:race).where('races.start_time <= ?', time)
-                               .order('races.start_time desc')
+  def second_last_race_order(results)
+    results.second&.order&.to_i || 0
   end
 end